refactor(frontend): extract shared PerformanceChart component
Create a shared PerformanceChart component to eliminate duplication between
ModelPerformanceStats and ChannelPerformanceStats.

Changes:
- Create performance-chart.tsx with shared chart logic (364 lines)
- Refactor model-performance-stats.tsx to 41 lines (was 327)
- Refactor channel-performance-stats.tsx to 35 lines (was 334)
- Export shared types: PerformanceDataPoint, PerformanceDisplayMode, LegendItem
- Both components now map data and delegate rendering to PerformanceChart

Benefits:
- ~580 lines of duplicated code eliminated
- Single source of truth for chart behavior
- Easier to maintain and extend
